+++ src/main/kotlin/io/jangseongbin/supplieswiki/theme/controller/ThemeController.kt
@@ -1,0 +1,4 @@
+package io.jangseongbin.supplieswiki.theme.controller
+
+class ThemeController {
+}
