+++ src/main/kotlin/io/jangseongbin/supplieswiki/theme/data/Content.kt
@@ -1,18 +1,16 @@
-package io.jangseongbin.supplieswiki.theme.domain
+package io.jangseongbin.supplieswiki.theme.data
 
 import jakarta.persistence.Embeddable
 import java.time.LocalDateTime
 import java.time.LocalDateTime.now
 import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.annotation.LastModifiedDate
 
 @Embeddable
-class Category(
-    val name: String,
+class Content(
+    val userId: Long,
+
+    val contents: String,
 
     @CreatedDate
     val createdAt: LocalDateTime = now(),
-
-    @LastModifiedDate
-    val updatedAt: LocalDateTime? = null,
 )
