+++ src/main/kotlin/io/jangseongbin/supplieswiki/theme/data/ThemeRepository.kt
@@ -1,0 +1,6 @@
+package io.jangseongbin.supplieswiki.theme.data
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface ThemeRepository : JpaRepository<Theme, Long> {
+}
