+++ src/main/kotlin/io/jangseongbin/supplieswiki/theme/request/CreateTheme.kt
@@ -1,0 +1,4 @@
+package io.jangseongbin.supplieswiki.theme.request
+
+class CreateTheme {
+}
