+++ src/main/kotlin/io/jangseongbin/supplieswiki/theme/service/ThemeService.kt
@@ -1,0 +1,4 @@
+package io.jangseongbin.supplieswiki.theme.service
+
+class ThemeService {
+}
